requirements/requirements-dashboard.txt
altair==5.5.0
pandas==2.3.3
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
streamlit==1.50.0
streamlit-autorefresh==1.0.1
streamlit-javascript==0.1.5